@@ -1,7 +1,7 @@
 package com.machinelearningforsmallbusiness.leetcodepython.utilities;
 
 import android.content.Context;
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 import android.os.AsyncTask;
 import android.os.Bundle;
 import android.util.Log;
