@@ -1,14 +1,12 @@
 package com.machinelearningforsmallbusiness.leetcodepython.utilities;
 
 import android.content.Context;
-import android.content.Intent;
-import android.os.Bundle;
-import android.support.v7.widget.RecyclerView;
+
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 import android.widget.ImageView;
 import android.widget.TextView;
 
-import com.machinelearningforsmallbusiness.leetcodepython.DisplayCodeActivity;
 import com.machinelearningforsmallbusiness.leetcodepython.R;
 
 import java.util.HashMap;
